fix codeclimate and leo review
